[APPAI-1434] Using connecxion library for introducing ingestion API
