+++ f8a_jobs/api_v1.py
@@ -23,6 +23,7 @@ from f8a_jobs.models import JobToken
 from f8a_jobs.defaults import AUTH_ORGANIZATION
 import f8a_jobs.defaults as configuration
 from f8a_jobs import graph_sync
+from f8a_jobs.graph_ingestion import ingest_epv_into_graph
 
 logger = logging.getLogger(__name__)
 
@@ -515,3 +516,8 @@ def invoke_graphsync_epv(**kwargs):
     """
     result = graph_sync.invoke_sync(params=kwargs)
     return result
+
+
+def ingest_epv(**kwargs):
+    """Trigger the EPV ingestion job based on the JSON sent via request."""
+    return ingest_epv_into_graph(kwargs, headers=request.headers)
