+++ requirements.in
@@ -7,7 +7,7 @@ Werkzeug<=0.16.1
 requests
 apscheduler
 beautifulsoup4
-connexion
+connexion[swagger-ui]
 Flask
 Flask-OAuthlib<=0.9.4
 Flask-Script
