+++ requirements.txt
@@ -11,8 +11,8 @@ apscheduler==3.6.3        # via -r requirements.in
 beautifulsoup4==4.9.1     # via -r requirements.in, f8a-worker
 billiard==3.6.3.0         # via celery
 blinker==1.4              # via raven
-boto3==1.14.48            # via -r requirements.in, f8a-worker
-botocore==1.17.48         # via -r requirements.in, boto3, f8a-worker, s3transfer
+boto3==1.14.51            # via -r requirements.in, f8a-worker
+botocore==1.17.51         # via -r requirements.in, boto3, f8a-worker, s3transfer
 celery==4.4.7             # via selinon
 certifi==2020.6.20        # via requests
 chardet==3.0.4            # via requests
@@ -21,7 +21,7 @@ clickclick==1.2.2         # via connexion
 codegen==1.0              # via selinon
 colorama==0.4.3           # via radon, rainbow-logging-handler
 configobj==5.0.6          # via anymarkup
-connexion==2.7.0          # via -r requirements.in
+connexion[swagger-ui]==2.7.0  # via -r requirements.in
 daiquiri==2.1.1           # via json2sql
 docutils==0.15.2          # via botocore
 git+https://github.com/fabric8-analytics/fabric8-analytics-utils.git@44c123b#egg=f8a_utils  # via -r requirements.in
@@ -41,7 +41,7 @@ idna==2.10                # via requests
 importlib-metadata==1.7.0  # via flake8, kombu
 inflection==0.5.1         # via connexion
 itsdangerous==1.1.0       # via flask
-jinja2==2.11.2            # via flask
+jinja2==2.11.2            # via flask, swagger-ui-bundle
 jmespath==0.10.0          # via boto3, botocore
 jsl==0.2.4                # via f8a-worker
 json2sql==1.0.0rc1        # via -r requirements.in
@@ -77,6 +77,7 @@ six==1.15.0               # via anymarkup-core, apscheduler, configobj, mando, o
 smmap==3.0.4              # via gitdb
 soupsieve==2.0.1          # via beautifulsoup4
 sqlalchemy==1.3.19        # via -r requirements.in, f8a-worker
+swagger-ui-bundle==0.0.8  # via connexion
 tenacity==6.2.0           # via f8a-worker
 toml==0.9.4               # via anymarkup, f8a-worker
 tzlocal==2.1              # via apscheduler
